latam-genai-demo/app/requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

db-dtypes==1.1.1
faiss-cpu==1.7.4
google-auth==2.27.0
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.14.0
Jinja2==3.1.2
langchain-community==0.0.17
langchain-core==0.1.18
langchain-google-vertexai==0.0.3
langchain==0.1.5
pandas==1.5.3
Pillow==9.5.0
plotly==5.18.0
pypdf==3.8.1
scikit-learn==1.2.2
streamlit-drawable-canvas==0.9.3
streamlit==1.27.2
youtube-transcript-api==0.6.1
